begin re-organization
