+++ bin/quilt-init
@@ -1,0 +1,1 @@
+#!/usr/bin/env node
