--- index.js
@@ -1,3 +1,0 @@
-exports.init = require('./tasks/init');
-exports.daemon = require('./tasks/daemon');
-exports.undaemon = require('./tasks/undaemon');
