--- tasks/daemon.js
@@ -1,7 +1,0 @@
-module.exports = function (mount, remote) {
-  if (!mount || !remote) {
-    throw new TypeError();
-  } else {
-    console.log('you did it!');
-  }
-};
