--- tasks/undaemon.js
@@ -1,3 +1,0 @@
-module.exports = function (mount, remote) {
-  console.log('you did it!');
-};
